+++ biojava-structure/src/main/java/org/biojava/nbio/structure/secstruc/DSSPParser.java
@@ -28,10 +28,8 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
 import java.io.StringReader;
-import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.zip.GZIPInputStream;
 
 import org.biojava.nbio.structure.Group;
 import org.biojava.nbio.structure.ResidueNumber;
@@ -100,32 +98,6 @@ public class DSSPParser {
 	}
 
 	/**
-	 * Fetch and parse the DSSP file of the specified pdb code
-	 * from the PDB web server and return the secondary structure
-	 * annotation as a List of {@link SecStrucState} objects.
-	 *
-	 * @param pdb path to the DSSP file to parse
-	 * @param structure Structure object associated to the dssp
-	 * @param assign assigns the SS to the structure if true
-	 * @return a List of SS annotation objects
-	 * @throws StructureException
-	 * @throws IOException
-	 */
-	public static List<SecStrucState> fetch(String pdb,
-			Structure structure, boolean assign)
-					throws IOException, StructureException {
-
-		URL url = new URL("http://files.rcsb.org/dssp/" +
-				pdb.toLowerCase().substring(1, 3) + "/" +
-				pdb.toLowerCase() + "/" +
-				pdb.toLowerCase() + ".dssp.gz");
-		InputStream in = new GZIPInputStream(url.openStream());
-		Reader read = new InputStreamReader(in);
-		BufferedReader reader = new BufferedReader(read);
-		return generalParse(reader, structure, assign);
-	}
-
-	/**
 	 * Parse a DSSP format String and return the secondary structure
 	 * annotation as a List of {@link SecStrucState} objects.
 	 *
@@ -183,16 +155,16 @@ public class DSSPParser {
 
 				//Parse the Bridge partners - TODO parallel or antiparallel?
 				String bp = line.substring(25,29).trim();
-				if (bp != "") {
+				if (!bp.equals("")) {
 					BetaBridge bb = new BetaBridge(
-							index, Integer.valueOf(bp), BridgeType.parallel);
+							index, Integer.parseInt(bp), BridgeType.parallel);
 					ss.addBridge(bb);
 				} else logger.warn("Unable to parse beta Bridge for resn "+index);
 
 				bp = line.substring(29,33).trim();
-				if (bp != "") {
+				if (!bp.equals("")) {
 					BetaBridge bb = new BetaBridge(
-							index, Integer.valueOf(bp), BridgeType.parallel);
+							index, Integer.parseInt(bp), BridgeType.parallel);
 					ss.addBridge(bb);
 				} else logger.warn("Unable to parse beta Bridge for resn "+index);
 
@@ -203,7 +175,7 @@ public class DSSPParser {
 					int b = a + 8;
 
 					String val = line.substring(a,b).trim();
-					if (val == "") {
+					if (val.equals("")) {
 						logger.warn("Unable to parse energy for resn "+index);
 						continue;
 					}
@@ -212,7 +184,7 @@ public class DSSPParser {
 
 					int partner = Integer.parseInt(p[0]);
 					if (partner != 0) partner += index;
-					double energy = Double.valueOf(p[1]) * 1000.0;
+					double energy = Double.parseDouble(p[1]) * 1000.0;
 
 					switch(i){
 					case 0:
@@ -236,15 +208,15 @@ public class DSSPParser {
 
 				//Angle properties
 				String val = line.substring(91,97).trim();
-				if (val != "") ss.setKappa(Float.valueOf(val));
+				if (!val.equals("")) ss.setKappa(Float.parseFloat(val));
 				else logger.warn("Unable to parse kappa for resn "+index);
 
 				val = line.substring(103,109).trim();
-				if (val != "") ss.setPhi(Float.valueOf(val));
+				if (!val.equals("")) ss.setPhi(Float.parseFloat(val));
 				else logger.warn("Unable to parse phi for resn "+index);
 
 				val = line.substring(109,116).trim();
-				if (val != "") ss.setPsi(Float.valueOf(val));
+				if (!val.equals("")) ss.setPsi(Float.parseFloat(val));
 				else logger.warn("Unable to parse psi for resn "+index);
 
 				if (assign) parent.setProperty(Group.SEC_STRUC, ss);